Move ScoredAnswer field/widget to to own include
